Exporting edges
Added the export of the parameters list
Added the possibility to export the (Leading) edges of the image. Still need to add the trailing edges ideally combining them in the same sheet with a double header (n line) on top and (L)(R) underneath.
Also need to link to the checkbox to require the export of the edges